Move FunctionToStringCutToZero to separate file
